devtools/client/shared/components/test/mochitest/test_tree_02.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test that collapsed subtrees aren't rendered.
-->
<head>
  <meta charset="utf-8">
  <title>Tree component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = async function () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");
    let Tree = React.createFactory(browserRequire("devtools/client/shared/components/VirtualizedTree"));

    const tree = ReactDOM.render(Tree(TEST_TREE_INTERFACE), window.document.body);

    isAccessibleTree(tree);
    TEST_TREE.expanded = new Set("MNO".split(""));
    await forceRender(tree);

    isRenderedTree(document.body.textContent, [
      "A:false",
      "M:false",
      "-N:false",
      "--O:false",
    ], "Collapsed subtrees shouldn't be rendered");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
};
</script>
</pre>
</body>
</html>
